Story #14025: fix cert generation scripts

+++ deployment/pki/scripts/lib/stores.sh
@@ -147,15 +147,18 @@ function generateTrustStore {
                     ${ALIAS}
     done
 
-    # Add the hosts/localhost certificates to the truststore
-    # for CRT_FILE in $(ls ${REPERTOIRE_CERTIFICAT}/server/hosts/localhost/*.crt); do
-    #     pki_logger "Ajout de ${CRT_FILE} dans le truststore ${CLIENT_TYPE}"
-    #     ALIAS="server-$(basename ${CRT_FILE})"
-    #     addCrtInJks ${JKS_TRUST_STORE} \
-    #                 ${TRUST_STORE_PASSWORD} \
-    #                 ${CRT_FILE} \
-    #                 ${ALIAS}
-    # done
+    if [ "${DEV_MODE}" == "true" ]; then
+        pki_logger "DEV_MODE is true"
+        # Add the hosts/localhost certificates to the truststore
+        for CRT_FILE in $(ls ${REPERTOIRE_CERTIFICAT}/server/hosts/localhost/*.crt); do
+            pki_logger "Ajout de ${CRT_FILE} dans le truststore ${CLIENT_TYPE}"
+            ALIAS="server-$(basename ${CRT_FILE})"
+            addCrtInJks ${JKS_TRUST_STORE} \
+                        ${TRUST_STORE_PASSWORD} \
+                        ${CRT_FILE} \
+                        ${ALIAS}
+        done
+    fi
 }
 
 function generateHostKeystore {
@@ -183,9 +186,11 @@ function generateHostKeystore {
                 ${P12_KEYSTORE} \
                 ${TMP_P12_PASSWORD}
 
-    if [ -f ${P12_KEYSTORE} ]; then
-        pki_logger " /!\ Suppression du p12: ${P12_KEYSTORE}"
-        rm -f ${P12_KEYSTORE}
+    if [ "${DEV_MODE}" != "true" ]; then
+        if [ -f ${P12_KEYSTORE} ]; then
+            pki_logger " /!\ Suppression du p12: ${P12_KEYSTORE}"
+            rm -f ${P12_KEYSTORE}
+        fi
     fi
 }
 
@@ -270,9 +275,11 @@ function main() {
             P12_KEYSTORE=${REPERTOIRE_KEYSTORES}/client-${CLIENT_TYPE}/keystore_${COMPONENT}.p12
             P12_PASSWORD=$(getKeystorePassphrase "keystores_client_${CLIENT_TYPE}_${COMPONENT}")
 
-            if [ -f ${P12_KEYSTORE} ]; then
-                pki_logger " /!\ Suppression du p12: ${P12_KEYSTORE}"
-                rm -f ${P12_KEYSTORE}
+            if [ "${DEV_MODE}" != "true" ]; then
+                if [ -f ${P12_KEYSTORE} ]; then
+                    pki_logger " /!\ Suppression du p12: ${P12_KEYSTORE}"
+                    rm -f ${P12_KEYSTORE}
+                fi
             fi
 
             pki_logger "Génération du p12"
